buttonfixed1

+++ VueReal/src/views/DangerView.vue
@@ -1,18 +1,34 @@
-<template >
+<template>
 
 
-    <div class="todo-container bg-success-subtle ">
-        <h1>This option is only if you want to delete the ENTIRE list
+  <div class="todo-container bg-success-subtle ">
+    <h1>This option is only if you want to delete the ENTIRE list
 
-        </h1> <button @click=deList>Delete all the list</button>
-    </div>
+    </h1> <button @click=dellList>Delete all the list</button>
+  </div>
 
 </template>
 
 <script setup>
 function deList() {
-localStorage.setItem("nameList", [])
+  localStorage.setItem("nameList", [])
 }
+
+
+function dellList(index) {
+
+    if (window.confirm("Do you really want to delete the list?")) {
+      localStorage.setItem("nameList", [])
+      saveListToLocalStorage();
+    } else {
+    nameList.value.splice(index, 1);
+    saveListToLocalStorage();
+  }
+}
+
+
+
+
 
 </script>
 
@@ -29,13 +45,13 @@ localStorage.setItem("nameList", [])
   border: 2px solid #000000;
   border-radius: 0px;
   padding: 0px;
-  
-margin: 50px;
+
+  margin: 50px;
   background-color: #f8f9fa;
 }
 
-.style-a{
- 
+.style-a {
+
   color: black !important;
   font-weight: bold;
   margin: auto;
@@ -44,6 +60,4 @@ margin: 50px;
 .bg-success-subtle {
   padding: 40px;
 }
-
-
 </style>
